refactor: 위치 이동으로 인한 변경
