+++ Attendance/src/main/java/com/example/Attendance/config/attendanceJob/AttendanceBatchJobListener.java
@@ -1,5 +1,6 @@
 package com.example.Attendance.config.attendanceJob;
 
+import com.example.Attendance.config.attendanceJob.step.attendance.SalaryBatchState;
 import com.example.Attendance.config.attendanceJob.step.email.EmailBatchState;
 import com.example.Attendance.config.attendanceJob.step.pdf.PdfBatchState;
 import lombok.RequiredArgsConstructor;
@@ -17,19 +18,16 @@ public class AttendanceBatchJobListener {
 
     private final PdfBatchState pdfBatchState;
     private final EmailBatchState emailBatchState;
-    private final BatchJobState batchJobState;
+    private final SalaryBatchState salaryBatchState;
 
     @Bean
     public JobExecutionListener attendanceJobListener() {
         return new JobExecutionListener() {
             @Override
             public void afterJob(JobExecution jobExecution) {
-                // 컨텍스트에 다시 저장
                 pdfBatchState.reset();
                 emailBatchState.reset();
-                batchJobState.reset();
-//                batchJobState.getFailedEmployee().stream().map();
-                // 사장에게 메일보내기 한번에
+                salaryBatchState.reset();
             }
         };
     }
