+++ Attendance/src/main/java/com/example/Attendance/config/attendanceJob/AttendanceJobConfig.java
@@ -1,14 +1,8 @@
 package com.example.Attendance.config.attendanceJob;
 
-import com.example.Attendance.config.attendanceJob.step.attendance.AttendanceBatchProcessor;
-import com.example.Attendance.config.attendanceJob.step.attendance.AttendanceBatchReader;
-import com.example.Attendance.config.attendanceJob.step.attendance.AttendanceBatchWriter;
-import com.example.Attendance.config.attendanceJob.step.email.EmailBatchProcessor;
-import com.example.Attendance.config.attendanceJob.step.email.EmailBatchReader;
-import com.example.Attendance.config.attendanceJob.step.email.EmailBatchWriter;
-import com.example.Attendance.config.attendanceJob.step.pdf.PdfBatchProcessor;
-import com.example.Attendance.config.attendanceJob.step.pdf.PdfBatchReader;
-import com.example.Attendance.config.attendanceJob.step.pdf.PdfBatchWriter;
+import com.example.Attendance.config.attendanceJob.step.attendance.SalaryBatchStep;
+import com.example.Attendance.config.attendanceJob.step.email.EmailBatchStep;
+import com.example.Attendance.config.attendanceJob.step.pdf.PdfBatchStep;
 import com.example.Attendance.dto.batch.BatchInputData;
 import com.example.Attendance.dto.batch.BatchOutputData;
 import com.example.Attendance.dto.batch.email.EmailInputData;
@@ -31,22 +25,13 @@ import org.springframework.transaction.PlatformTransactionManager;
 @RequiredArgsConstructor
 public class AttendanceJobConfig {
 
-    private final AttendanceBatchJobListener attendanceBatchJobListener;
-
-    private final AttendanceBatchProcessor attendanceBatchProcessor;
-    private final AttendanceBatchWriter attendanceBatchWriter;
-    private final AttendanceBatchReader attendanceBatchReader;
-
-    private final EmailBatchProcessor emailBatchProcessor;
-    private final EmailBatchReader emailBatchReader;
-    private final EmailBatchWriter emailBatchWriter;
-
-    private final PdfBatchProcessor pdfBatchProcessor;
-    private final PdfBatchReader pdfBatchReader;
-    private final PdfBatchWriter pdfBatchWriter;
+    private final SalaryBatchStep salaryBatchStep;
+    private final EmailBatchStep emailBatchStep;
+    private final PdfBatchStep pdfBatchStep;
 
     private final JobRepository jobRepository;
     private final PlatformTransactionManager transactionManager;
+    private final AttendanceBatchJobListener attendanceBatchJobListener;
 
 
     @Bean
@@ -63,9 +48,9 @@ public class AttendanceJobConfig {
     public Step attendanceStep() {
         return new StepBuilder("automaticTransferStep", jobRepository)
                 .<BatchInputData, BatchOutputData>chunk(3, transactionManager)
-                .reader(attendanceBatchReader.attendanceReader())       // 데이터 읽기
-                .processor(attendanceBatchProcessor.attendanceProcessor()) // 데이터 처리
-                .writer(attendanceBatchWriter.attendanceWriter())       // 데이터 쓰기
+                .reader(salaryBatchStep.salaryReader())       // 데이터 읽기
+                .processor(salaryBatchStep.salaryProcessor()) // 데이터 처리
+                .writer(salaryBatchStep.salaryWriter())       // 데이터 쓰기
                 .build();
     }
 
@@ -73,9 +58,9 @@ public class AttendanceJobConfig {
     public Step statementPdfStep() {
         return new StepBuilder("statementPdfStep", jobRepository)
                 .<PdfInputData, PdfOutputData>chunk(3, transactionManager)
-                .reader(pdfBatchReader.pdfReader())       // 데이터 읽기
-                .processor(pdfBatchProcessor.pdfProcessor()) // 데이터 처리
-                .writer(pdfBatchWriter.pdfWriter())       // 데이터 쓰기
+                .reader(pdfBatchStep.pdfReader())       // 데이터 읽기
+                .processor(pdfBatchStep.pdfProcessor()) // 데이터 처리
+                .writer(pdfBatchStep.pdfWriter())       // 데이터 쓰기
                 .build();
     }
 
@@ -83,9 +68,9 @@ public class AttendanceJobConfig {
     public Step statementEmailStep() {
         return new StepBuilder("statementEmailStep", jobRepository)
                 .<EmailInputData, EmailOutputData>chunk(3, transactionManager)
-                .reader(emailBatchReader.emailReader())       // 데이터 읽기
-                .processor(emailBatchProcessor.emailProcessor()) // 데이터 처리
-                .writer(emailBatchWriter.emailWriter())       // 데이터 쓰기
+                .reader(emailBatchStep.emailReader())       // 데이터 읽기
+                .processor(emailBatchStep.emailProcessor()) // 데이터 처리
+                .writer(emailBatchStep.emailWriter())       // 데이터 쓰기
                 .build();
     }
 }
